ezcommon-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# AWS SDK
boto3==1.35.36
botocore==1.35.36

# Authentication and security
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# OpenSearch
opensearch-py==2.7.1
requests-aws4auth==1.3.1

# ChromaDB (optional database backend)
chromadb>=0.4.0

# PDF processing
PyPDF2==3.0.1

# PDF to image conversion for Vision API
pdf2image==1.17.0
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1

# OpenAI for chatbot and vision
openai>=1.99.5

# Gemini for alternative LLM
google-generativeai>=0.8.0

# HTTP client
requests==2.32.3

# Validation
email-validator
